requirements.txt
opencv-python==4.9.0.80
mediapipe-silicon==0.9.2.1
streamlit==1.30.0
streamlit-webrtc==0.55.0
av==14.0.0